requirements.txt
ipykernel
langchain-community==0.3.27
langchain-openai==0.3.29
langchain==0.3.27
langgraph==0.6.4
nbformat==5.10.4
openai==1.99.9
openpyxl==3.1.0
pandas==2.2.2
Pillow==10.0.0
plotly==5.24.1
python-dotenv==1.1.1
seaborn==0.13.2
streamlit_option_menu==0.4.0
streamlit-extras==0.7.1
streamlit==1.45.1
watchdog==6.0.0